Add test cases for prerelease version regex
